partial stuff

+++ packages/drummachine/lib/screens/base_screen.dart
@@ -21,10 +21,10 @@ class BaseScreen implements GameScreen {
   void logicLoop() {}
 
   @override
-  void mouseClick(double x, double y, MouseButton mouseButton) {}
+  void mouseClick(int x, int y, MouseButton mouseButton) {}
 
   @override
-  void mouseMove(double x, double y) {}
+  void mouseMove(int x, int y) {}
 
   @override
   void onEnter() {
