+++ packages/teenytinytwodee/lib/application/game_screen.dart
@@ -16,7 +16,7 @@ abstract class GameScreen {
 
   void keyboard(int keyCode);
 
-  void mouseClick(double x, double y, MouseButton mouseButton);
+  void mouseClick(int x, int y, MouseButton mouseButton);
 
-  void mouseMove(double x, double y);
+  void mouseMove(int x, int y);
 }
